BIOL503_Lab5_Rscripts
BIOL503_Project_FilterData_Rscript_2025-02-10.R is the script used to filter the raw data and split the metadata column. Code to rarefy the data is also present in this code but there are extra comments still present from lab 2. Feel free to delete and edit then push a new file.

BIOL503_Lab5_TaxaBarChart_Rscript_2025-02-10.R is the script used to create stacked bar chart of relative abundance at the genus level for just the root microbial community of the wash treatment.