Add files via upload
I am struggling to get the taxa plot to work, im trying to follow the disease code, as it creates others per group. Idk whats going wrong, but i think its to do with calculating relative abundance. the disease code has so many groupings but we dont so its likely a translation somewhere in there, as we dont subset as much?